+++ src/main/java/com/reviewer/entity/ReviewSummary.java
@@ -9,6 +9,7 @@ import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.mongodb.core.index.Indexed;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
@@ -31,5 +32,5 @@ public class ReviewSummary {
 
     private Long totalReviews;
 
-    private Float average;
+    private BigDecimal average;
 }
